Reworked file IO of blocks.
Fixed bugs causing failure to read in any blocks. Also completely
upgraded how the XML structure of blocks.

The type is now an attribute within the block's starting element. The
name of elements is now the name of the type of block; as a result added
a new virtual function for that within block factories.