Worked on block model/view system.
Primarily worked on block view class but also made small changes to main
window and block model that interact with it. Worked out how blocks are
selected with block view class and how that is communicated to the main
window. Fixed bug in block model index function that was causing crashes
along with adding missing keywords and removing unused functions.