Refactored code format.
Got rid of silly comments that did not add any meaningful description.
This will be part of this reference program!

Also changed name to pointers that an object takes ownership of to begin
with taken instead of o_.

As a result of removing all the silly comments, added additional
functions from single functions to make it more readable.