Added C++/Qt template type.
As a result made a number of changes to the abstract type interface
needed for the template type to work properly. Most notably the scope
list needed to be past to the template name virtual function for type
values within the template.

Also added equality operators to abstract type which tests scope and
isEquivalent within it.

Lastly, added an exists checker function to the container class that
will check to make sure a given type exists within the container. This
will be used later once types are used by variables/functions.