Worked on entire project.
Changed qlist arguments to references where appropriate. Made the C++/Qt
block factory a singleton class so there is no need for a global object.

Lastly, Reworked abstract block to have type and factory getters as
virtual functions that implementations are responsible for doing now.
This makes all constructors the most basic and simple and lessens the
chance for errors.