Added title bar to block view.
As a result also changed the block view types to reflect their name is
already presented now as part of the block view. Did this to minimize
code as any block will display their name first as a title. Also added
their respective icons before the title to make it look good.