Added C++/Qt type edit widget.
As a result made farily large changes to the entire type system for C++/
Qt. It is not integrated with the definition block type and works as
intended.

Lots of debugging with the entire type system from integrating the new
edit widget which allowed for testing of the read/write abilities of
types. All bugs found and everything works for C++/Qt type system now.