Modified C++/Qt type interface.
Now has additional save functionality to remember its scope and depth
used to output its scoped name. Depth has also been added which tells it
how many namespaces deep it should report its scope if any or simply add
all namespaces. As a result of all this the fullName() function was
changed to not include scope and the scopedName() function which changed
in the same way.