Removed C++/Qt type system.
After much thought removed entire system. It will be replaced by simple
strings stored in namespace blocks instead. This decision was made
because the type system was becoming very complex and really didn't fit
the purpose of this program. Just in case all that work was saved on an
alternative type branch.