Fixed signalling bug.
Found bug that caused segfaults because the block view was deleting the
view class that was still being used after the signal that caused its
deletion; fixed. As a result have to rethink the bodyChanged() signal
which is now removed from abstract block as it does not belong there.
Will add signal to function view class and have it change itself.