Refactored C++/Qt function classes.
Redid flags to be called properties and changed their setters and getters
to booleans for each specific property. As a result added different icons
for virtual and abstract functions. As a result of that added an empty
class type to factory and also added to the block view to update the
actions and menu if the data model currently selected has changed.