Refactored block interface.
Moved the creation of view and edit blocks to the abstract blocks
themselves instead of the block factory as this makes much more sense.
It ironically revealed a const bug where the view could have potentially
modified the block and as a result all views were changed to hold only
constant references.