Refactored abstract blocks.
Changed how making copies of an abstract block works. Now the abstract
class makes the copy and calls two new interface functions for
initializing to new copy and then copying all of its data. This was done
because this pattern ended up happening in every single block
implementation anyway. The makeCopy() function is still virtual and can
be overridden if an implementation wants to do something different.