Added C++/Qt signal blocks.
As a result found and fixed two bugs. First bug caused segfault when
function edit was attempting to query checkboxes that did not exist.
Second bug was class not updating based off of functions being virtual
or abstract; this was caused because of the selection methods in access
being to specific along with the child update methods in the class block
being changed to remove it looking for that type of change.