Added to C++/Qt blocks.
Forgot to add inheritance string for classes. Also forgot to add
namespaces to type checking so before if you used :: for scoping it
would say the type is invalid. Both items were added.

As a result realized the applyClicked() function for abstract edit
needed to be changed to return a boolean informing if the apply was
successful or not. If it is not successful the edit widget does not
close if ok is clicked. This was needed for the class inheritance to
inform the user if the inheritance string is invalid and have abstract
edit leave itself open.