Modified base block system.
Made many final modifications and additions to the base block system.
Changed model for blocks to be more intuitive along with chaning
blockview class to be the same. Added shortcut keys for the edit menu
along with adding the ability to bring it up as a context menu when
right clicking the tree view.

Also fixed a segfault bug in C++/Qt access edit caused by not
initualizing pointers before using them.