Added C++/Qt global parser.
As a result had to change basic parsing system as it was forgotten that
parsers will usually need a pointer to he root block of a project. Also
added some additional functinality to the abstract parser for adding
lines, giving a default of adding an empty line and also specifying an
indent of spaces to be prepended to the line.