Worked on C++/Qt parsing.
Worked out a multitude of bugs causing output issues with parsing
classes mostly focused on the header class. Also fixed a couple bugs in
the block system, such as not being able to edit the name of signals or
slots.

Added a new access parser for adding access markers to class definitions.
Also added an empty source parser that will be for parsing source files.