Fixed C++/Qt parsing bug.
Fixed bug where the output of common header files would include the
#endif of the header guard as custom header code and keep appending an
additional #endif at the beginning each time the header was scanned.

As a result a new special barrier line had to be made, an empty single
line comment "//" denotes the end of the header section of a source/
header file.