Added file version system.
Finished adding version system for abstract blocks when reading from or
writing to XML. This now allows for updating how different blocks are
saved while being able to open older versions of socrates reference
files.

Also added a new version for all blocks which fixes the attributes bug
where the order of attributes in XML is not promised resulting in the
file dramatically changing everytime it was saved even if no actual
changes were made. This was a serious bug because this program is meant
to have its files saved in version control systems.