Modified presentation of C++/Qt blocks.
Changed presentation of functions and variables in a major way. Now only
their names are fully displayed in their tree view. They have special
markers now for things like return variables or initializers. The
details of that information is not published in their respective views.
For functions the arguments are listed first because that partially
identifies them in C++.