Fixed C++/Qt template parsing bug.
Fixed bug in regard to templates where they would not output their type
in template declarations. This was caused because of how the variable
block was changed that it inherits.

As a result of fixing the bug removed deprecated common C++/Qt functions,
moving one to the only class where it was used.