Refactored C++/Qt edits/views.
Changed all edits and views of blocks to no longer use casting since
their blocks actually make them now. This was residual from when that
was not the case.

As a result also cleaned up both edits and views to a large degree. Most
views now simply have headers and nothing else.