Fixed C++/Qt parsing bug.
Fixed a bug that caused functions nested with extra spaces to not be
correctly parsed in multiple ways. The first was was that the opening and
closing brackets were not read correctly because they contained spaces.
The second and much more difficult to fix was was the extra spaces not
being accounted for when reading in the lines of code in the function.

As a result added the constant flag to an abstract parser method.