Modified block system.
Fixed a serious design bug concerning the model/view system. The patch
work code that was in place for handling changes to the body of a block
was bypassing the data model system of the block model class. This was
fixed by adding custom data roles for names and bodies of blocks, along
with removing the adhoc patch used in the abstract edit class. This
required updating the abstract block class itself and all block classes
to use the new system.