Fixed multiple bugs.
Fixed a segmentation fault bug in block view caused by cutting a block
if there was already a block in the copy buffer. Also fixed other minor
bugs in block view where the current index and context menu were not
upating properly when blocks are added/removed.

Fixed a general bug in C++/Qt operator block where it was causing the
notify name change signal to emit before it had a parent, causing
projects that contained any operators to throw an error when being
opened.