Refactored operation comments.
Changed function operation comments to be inline. The entire code base
now needs to be updated as the inline comments more than likely do not
make any sense right now.

Updated the inline comments for the singleton and singleton factory
classes. As a result modified the function parser code to not add a
number to the inline comments because it looks nicer.