Added custom dictionary spelling.
Finished adding a custom dictionary system that allows for adding
special spelling words on a per project basis. This allows for custom
words which is very common in programming projects.

As a result made other small changes to other classes.

The new dictionary interface is integrated in the text edit widget and
other associated classes, allowing the addition of new custom words
through its spell checking dialog. A new dictionary dialog is also added
to the main window to allow the user to view and remove custom words.