Fixed null blocks issue.
Resolved issue with all C++/Qt blocks being in an invalid null state
when the user creates a new one. As a result had to add an additional
argument to the abstract block factory interface for differentiating
between a new default block being created by the user and a null state
block being created while reading in a project.