Reverted abstract block interface.
Modified abstract block interface back to its more simple design before
the v0.5 change. Realized this was a mistake, and the field system is now
being replaced with the basic block system which will make it much
easier to define blocks.

As a result of changing the abstract block interface the C++/Qt project
type is not completely broken and will be refactored to using the basic
block system once it is complete.