Added basic block edit.
Almost finished all code for the editor class of the basic block class.
As a result ran into a problem when it came to defining variable types.
This needed to be moved to the basic block interface so type edit
widgets can be made in its editor. Managed to fix this by adding a new
interface in the basic block class that returns the list of available
types.

Also added a check function to basic block to trim down on repeated
code.