Finished basic block editor.
Finished adding comments to the new basic block editor class. As a
result realized the new editor was missing the splitter layout intended
so had to add new methods and variables for that.

Also fixed the persistent dialog class to use a qt string for its key to
make it more versatile.

Also made some slight cosmetic changes to the basic block and abstract
block classes.