Worked on C++/Qt blocks.
Finished integrating type list and type blocks for given project type.
As a result realized the basic block view interface for adding custom
elements did not work because it has to be called after construction of
the instance; this was fixed.

Realized the qt meta object system can be used for making new instances
of basic blocks so removed the make basic block interface in favor of
using that and not reinventing the wheel. The default constructor of any
basic block implementation must be invokable by the qt meta object
system for this to work.

Removed all old source and header files for the updated blocks that are
no longer used in the new basic block system.