Integrated C++/Qt blocks.
Finished adding function and template blocks into the new basic block
system. Was just going to add function block but realized template
blocks needed to be added  first. Many methods are still incomplete for
the function block that requires class and access blocks to be working
to finish.

Added a new combo box edit widget for the basic block edit system. Was
going ot be used for function blocks but ended up not needing it. Will
need it for access blocks so that is fine.

Added a base block back that now provides methods that provide the base
name and description fields that almost all other blocks use.