Added GLSL parsers.
Added variable and block parsers for GLSL along with a base parser class
that will be used for all GLSL parsers.

As a result added a new helper method in the abstract parser class for
adding generic comments to code. Also made some minor changes to the
GLSL parser factory.