Modified GLSL parsing.
Changed the GLSL root shader parser to be able to dynamically mimic the
order of blocks in the project with the order they are presented in the
source file. In other words they both follow the same order now, minus
the special main function. The special main function is still the last
object in the source file to the defined.