Worked on C++/Qt parsing.
Added wrapper classes for all relevant basic parsers. Added a lot of
functions within the parse namespace for building the scanner objects of
an entire project. Tested on a global header file and most things seem to
work. A lot of work and more functions are required until the new
parsing system for this project type is complete.

Also fixed some minor bugs found in the basic parsers.