Worked on C++/Qt parsing.
Finished the function for building the header file of namespaces in the
parsing namespace. As a result added to the basic function parser to use
a predefined header instead of the one that is matched with the given
input line. Also fixed many small bugs while testing the namespace header
function.