Worked on C++/Qt parsing.
Added ability to parse header files for classes. As a result had to
refactor a lot of code inside the parsing namespace to better work with
how classes work. Also added a set indent virtual method for all
abstract parsers as this was needed for definitions in parsing between
classes and their definitions.

Added a has any templates testing function to the variable block as that
was needed for determining if a variable required a definition or not.