Optimized scanner.
Optimized the find parser method in the scanner class which was causing
severe lag when scanning source files. This was because it constructed a
regular expression class for each input parser for each line for each
input file. Now the optimized method creates a single regular expression
only once and uses it to match against all input parsers.