Refactored exception handling and fixed bugs.
REVISION

Removed custom exception class and replaced with socrates utility
library io error exceptions and the standard library exceptions.

Fixed missing checks needed in project constructer for opening an
existing project to make sure it is a valid XML file and also a valid
project XML file. Appropriate read error exceptions are now thrown if
either is not the case.