*** Started python refactoring.
Began massive process of refactoring this application as a python program using
pyside2. Made initial socref package with some basic modules and classes such as
Block_Factory, Abstract_Block, and Project_Model. Also added initial package that
will be the python language. Languages will be added as a module that is
imported. Project types are no longer called that, instead they are called what
they are; a language.