* Added to block system.
Added a lot of functionality to the abstract block, block factory, and project
model classes. The project model can now use abstract blocks for its model. The
block factory allows the registration of a special root block class used for
creating the root block of a new project.

Modified the basic formatting of code in all other python files. This is still in
flux as a standard is figured out.