* Modified undo/redo in projects.
project_commands:
Removed qt undo command as a parent class of the command classes as it is not
needed and brings in an unwanted qt widgets dependency.

project_model:
Refactored the model to use its own internal undo command stack, removing the qt
widget dependency on the qt undo stack class.