* Modified parsing.
Changed how the build process of parsing works. Now instead of building to a
string, a list of lines are built. The two helper functions in the utility
module was updated to return a list of line strings instead of a string. All
python code was updated to build a list of code lines and only builds it into
file contents at the very end in the parser. This should make things much
faster. C++ has not been updated and as a result is currently broken.