* Added to python parsing.
socref_python.block.Class:
Added new "in file" flag for classes. If a class is in file in means its
definition is within its module file. If it is not in file then it is contained
in its own special container module and then imported to its namespace module
file. This allows flexibility and better support for python scripts.

socref_python.parser.Parser:
Added code to recognize new "in file" flag for the class block.

socref_python.block.Package:
Added code to recognize new "in file" flag for the class block.

Updated various module comments.