* Refactored python blocks.
Added many additional icons for methods and members depending on their access.
Deprecated the access block instead of removing it right away. Removing it right
away could cause other reference files to break if they have that block type. It
will be removed in the future. The default name for a python access block is now
DEPRECATED so the user knows it will be removed in the future.