* Added to python parsing.
Added ability for python parser to pick up scanned lines of code located at the
beginning of classes after their initial document string. Also added to the
parser's unknown interface so it returns any scanned lines of class code that
have not been used when building back the source code.

Building the scanned lines of code back into the class has not yet been added
because enumerations will have to be removed at the same time. This will be the
next and final step for adding support for code being recognized at the
beginning of classes.