*** Removed python enums.
Removed the enumeration and enumeration value block types for python because
python does not have a formal enumeration type. Enumerations are simply classes
in python and that is now how they should be defined in any reference using the
new support for adding lines of code to classes themselves.

Also removed all variable object definitions from the reference, instead
changing them to lines of code for the classes where they are defined. The only
objects that should now be defined in the reference for classes are Qt signals.