* Refactored code.
Moved all global object definitions in modules to script code lines instead of
being defined in the reference as global definitions. This must be done because
the new parser picks up object definitions in modules as a script line now. This
is fine since this feels much more python than it was previously.

Also fixed a few bugs in the new script parsing code to handle descriptors,
remove the special __main__ line, and remove the deprecated special script
comment line.