* Modified code organization.
Moved all public interface functions of the core application to a new module
called public. This includes everything in the now deleted edit module along
with the register function that was in the abstract module. This update follows
the new standard where everything that is public in a package should be exposed
in the public module, with the exception of abstract classes that are still
exposed in the abstract module.

Updated all language packages to use the new module structure of the core
application.