** Added to python parsing.
Added ability for python to not output a doc string for a function whose
description field is empty. This can be used for methods that inherit from other
methods and can use the same documentation.

Changed indent of closing parenthesis of function definition line to match the
first line if it is multi line. This makes the code more readable.