* Modified core application.
Made complete overhaul of the block factory singleton class. It is much more
modular now and does not deal with importing the package of a language itself.
Instead it now provides begin and end registration methods for block
registration of an imported language. Actually loading a language with imports
will now be done in a new yet to be done controller class that uses the block
factory.

Updated deprecated comments in the project model class.