** Added to core interface.
Finished initial rough coding for the new abstract parser and reader classes.
Comments are still missing or deprecated.

Added initial rough draft of abstract writer class. Comments are still missing.

This completes the first pass at making the new parser/reader/writer system for
parsing of a project's code files. This should make things much easier for
languages that implement it by making it much more modular than it was before.
Now instead of everything passing through one single parser instance, scanning
and building is delegated to as many abstract reader and writer classes as a
language wishes to make. In addition to this modular delegation a system for
looking up readers from writers is implemented to allow for easy organization of
anything needed by writers that readers scanned.