* Integrated new core interface.
Added new parser factory class that acts to create parsers, removing it as an
abstract method in the block class. Now when a parser is initialized it is
passed the root block as an argument. Renamed parser model to parse controller
and moved it to the controller namespace.

Finished integrating the new abstract classes to work within the core
application. This did not take much work since the abstract parser did not
change much when using it.

Added back the forgotten unknown method in the abstract parser class. Added
unknown interface in the abstract reader class which now simply returns any
lines of code for that specific reader not gotten by a writer during the build
process.

Fixed random lines of code here and there that was missed before the massive
overhaul of the code base to the new standard.