** Updated python language.
Finished initial update and integration of the python language package to use
the new core application interface with readers and writers. The language is
still very incomplete, right now only implementing a single block, reader, and
writer. This is so the interface can be tested and debugged without going any
farther. Moved the language's data files to the new location within the root
data directory.

Added additional method in abstract parser and reader classes to allow
discarding a saved cursor. Added additional method in abstract writer to take a
list of base keys added to the beginning of any lookup call.