* Added comments.
Finished adding comments to abstract reader class.

Added comments to abstract writer class, all edit definition classes, and all
output classes.

As a result of commenting the rich text output class, realized the order of
arguments for a couple methods were in the wrong order. Changed them to the
right order and updated their use in the python language implementation.