** Updated GLSL language.
Updated bare skeleton version of the GLSL language to integrate with the new
core application interface. The bare skeleton only implements a single root
program block with no readers or writers. A blank dummy parser is also
implemented so the language will import successfully. Moved the compiled
resources into the new data sub folder and out of the package folder.