* Modified abstract parser.
Modified private read all method of abstract parser to be able to handle a path
where nothing is returned by the reader interface for it. In this case that path
will be ignored for reading and it will move on to the next path. This is for
cases where source files exist that require no parsing for its output.