** Added to GLSL language.
Added variable block, variable writer, structure block, and structure writer
classes to the GLSL language implementation. Updated the shader writer to
initialize its children variable and structure writers.

Added key method to program block to allow for lookup of function readers in the
future once the reader is implemented. The key method required implementation in
the program block because it will be used by shader blocks in order to lookup
the special main function reader by the shader writer.

Modified comments of various GLSL and python writers so it is easier to
understand.

Modified root make file so each sub directory is on a separate line for easier
readability.