* Updated comments.
Updated comments for all classes in the abstract, base, edit, and output sub
packages of the core application. Updated comments for all controller, factory,
and GUI controller classes in the core application. Updated comments in the
decorator functions of the core application

Fixed harmless bug in the GUI highlight controller class in the core
application. This bug involved an extra argument in the initialization operator
that was no longer used in the class itself. Fixed by removing the deprecated
argument.