* Updated comments.
Updated comments for all classes and functions in the block and reader packages
of the python language implementation.

Updated comments for boolean flag arguments in the core application. The
arguments are now described in the description like all other arguments.

Fixed bad naming of variable named speller to spellCheck for multiple classes in
the core application.