* Updated function writer.
Updated the function writer class of the python implementation. When outputting
the document string for arguments, arguments with empty descriptions are ignored
and not written to the document string output. If All arguments are ignored then
the arguments title in the document string is not written.

Updated arguments for the writer base, parser factory, and project model classes
of the core application. Arguments that do not need document string output have
now been removed.

Fixed argument for the function block class of t he python implementation. The
flag argument for the decorator method was never actually documented and the
entire document string of the method was never updated. Fixed both issues by
updating all comments along with adding an actual description of the argument.