* Updated core classes.
Updated the project view class of the core application. Added an enumeration for
insertion methods. Updated all code within the class to use the new enumeration
class. Renamed the insert values private method to insert location because that
is a lot more descriptive and its original name was misleading because it never
actually inserted anything.

Updated the project model class of the core application. Added shorthand
assignment for the Role enumeration so the self reference is not needed when
used within the class code.