* Updated language implementations.
Added new access, class, enumeration, and variable block classes for the C++
implementation. Moved all appropriate code from the old block classes into the
new ones, updating code where required. Completely revamped how flags are
handled in the variable block to make it much more clean and python.

Renamed the namespace block to its correct name for the C++ implementation.

Updated the shader block class of the GLSL implementation. Completely revamped
how flags are handled identical to how it was done for the variable block class
in the C++ implementation.

Updated enumeration comments for the project view class in the core application.

Fixed bug in the block edit dock class of the core application. When building a
combo box the text and icon values were in the incorrect order from the provided
tuple.