** Updated C++ implementation.
Added a large amount of new icons for function blocks in the C++ implementation.
New icons for public, protected, and private access was created for every
possible function icon type. Removed deprecated icons for the access block that
is no longer needed.

Removed the access block class from the C++ implementation.

Updated the function block class in the C++ implementation. Added a new access
field that represents the access type for the function. This replaces the
removed access block.

Added the settings class for the C++ implementation. Moved all settings from the
old module to this new class, deleting the now deprecated module.

Added the parser class for the C++ implementation. For now this is an empty
parser implementation that creates an empty path list.

Added the C++ implementation back into the core application as an active
language. Fixed multiple bugs in the C++ implementation that was found after
adding back into the application.