*** Updated core application.
Updated the parser base class of the core application. Added an additional tuple
item to the list of path list tuples returned from the path list interface. This
additional item is an options dictionary that can pass additional information
required when making the path's reader and writer. Added the new options
argument to the reader and writer interfaces.

Updated all language implementation parsers so use the new options argument.