*** Updated core application.
Updated the parser base class of the core application. Removed the reader and
writer interfaces. Modified the path list interface to return tuples of relative
paths, optional abstract reader, and required abstract writer in that order.
This removes the need of the removed reader and writer interfaces, and allows an
implementation to have more control over the specific reader and writer objects
used for each source file in the generated path list. This is a much more
elegant way to make this new system compatible with the needs of C++ instead of
the previous workaround of adding an options argument.

Updated the writer base class of the core application. Added a link interface.
This is called to allow the writer to link itself to any required reader. This
used to be done in the constructor. This no longer works because the new parser
system creates the writer instances before the readers are called and added to
the shared parser instance's lookup table. The new link interface however is
called after all reader instances are called.

Updated all parsers and writers of all language implementations to use the new
path list and link interface system.