* Updated C++ implementation.
Added new head reader, source reader, head writer, and source writer classes for
the C++ implementation. These readers and writers deal with the main header and
source files of a C++ project, and build the foundation of all other readers and
writers.

Updated the namespace block class in the C++ implementation. Added a scope
method that returns a list of namespaces representing the scope of any block.
Added a key method that provides a unique key used for linking writers with
readers using their respective block.

Fixed minor grammar typo in comment of the module reader class in the python
implementation.