* Updated C++ implementation.
Added new namespace writer class for the C++ implementation. This writer outputs
the code for its corresponding namespace block as a header declaration. Needs to
be linked with a to be coded namespace reader.

Fixed comment typos in the head writer and source writer classes of the C++
implementation.

Fixed comment typo in the module writer class of the python implementation.