* Updated C++ implementation.
Added the C++ reader base class for the C++ implementation. This base reader
class holds all common parsing methods used by multiple readers.

Added the namespace reader class for the C++ implementation. This parses the
declaration of a namespace in its header file.

Updated the head and source reader classes in the C++ implementation. Updated
them to inherit from the new C++ reader base class. Updated all code to use the
common scan methods found in the new C++ reader base class. Added
implementations of the new scope getter method to provide proper scope to
children readers.

Updated the head, source, and namespace writer classes in the C++
implementation. Removed the namespace scoping from the namespace writer and
added it to both the head and source writers. Added the current namespace of the
source writer for namespaces to its scoping output because that is different
from the header declaration and only contains functions.

Updated description of base package to be more generic and accurate in the core
application.