* Updated C++ implementation.
Added new enumeration, union, and function writer classes for the C++
implementation. The enumeration and union writers are basic declaration writers.
The function writer is very complex, taking an extra argument for being a writer
for a function's declaration or definition.

Updated the function block class in the C++ implementation. Added getter methods
for information required by the new function writer that makes sense to be
contained within the block instead of the writer. Added new interface methods
for access types and determining if a function block has any definition to
write. These interfaces will be used when implementing the C++/Qt implementation
as an extension of the C++ implementation.

Updated the class and name space writer classes in the C++ implementation. Added
code for initializing children writers for function, enumeration, and union
children of the class or name space block of the writer.

Updated the source writer class in the C++ implementation. Added code for
initializing children writers for function children of the class or name space
block of the writer that should have their definition in the source writer's
source file. This requires special logic and recursively going down into deeper
children, with additional logic to prevent grabbing functions that do not
belong.