** Updated C++ implementation.
Added new function reader class for the C++ implementation.

Updated the block reader base class in the C++ implementation. Modified the
initialization name argument to be optional which causes this class to not set
its lookup key. This allows this class to be compatible as a parent for the
function reader since it needs to set its lookup key with its argument
signature.

Updated the file reader base class in the C++ implementation. Removed the
initialization block argument and code for conditionally adding the block's name
to the scope if it is a name space. This was causing problems and belongs in the
head reader.

Updated the function block class in the C++ implementation. Added number of
arguments to the display name. Added custom key method that removes the file
flag and adds the function's argument signature to the returned key.

Updated the name space block class in the C++ implementation. Added special
condition check to key method for functions, using its special name method
instead of the raw name property.

Updated the head reader class in the C++ implementation. Updated arguments
passed to changed parent class. Extended the scope method to add the reader's
block's name if the block is a name space and not the root name space.

Updated the source reader class in the C++ implementation. Updated the arguments
passed to changed parent class. Added scanning of function definitions, creating
new function reader children for each function definition found.

Updated the function writer class in the C++ implementation. Made a large amount
of small changes to the output of functions, mainly dealing with their
definitions. Changed name of some local objects to match the naming convention
of the new function reader class. Implemented the link interface to link with a
writer's function reader.

Updated the head and name space writer classes in the C++ implementation.
Modified and added where blank lines are written to output.

Finished updating entire C++ implementation.

Removed all old deprecated C++ python files that are no longer used.