* Updated C++ implementation.
Updated the block writer base class. Added new getter interface for returning
lines of code that come before the main lines of code enclosed in the bracket
block. This new interface is required for class writers to use this base class.
Added custom output if the writer's block is a class and has parent classes.

Updated the C++ reader base class. Updated the regular expression for finding
classes to account for a possible colon at the end if a class has parents.

Updated the class block class. Renamed the class parents method to the more
generic parents method.

Updated the class reader class. Modified the scanning interface to account for
classes that have parents, skipping the parent lines until the first opening
bracket is encountered.

Updated the class writer class. Implemented the pre lines interface to add
parent declarations for a class that has parents.