** Updated C++ implementation.
Updated the block writer base class. Removed name argument from initialization
operator and added interface to return the same value. This allows for
implementations to have more dynamic control over what name to return when the
block is generated.

Updated the C++ reader base class. Updated all regular expression matching to
remove public access statements that are no longer on the same line. Edited
enumeration regular expression to match C style enumerations.

Updated the enumeration block class. Added class flag that determines if the
enumeration is a class or not. This allows for C style enumerations when
required instead of only class enumerations.

Updated the class and union writer classes. Adapted to changes in parent block
writer base class.

Updated the enumeration writer class. Adapted to changes in parent block writer
base class and added conditional for name depending on of it is a class
enumeration or not.